Propagating sigint to subprocesses like terraform instead of killing
So I figured out why we kept getting terraform lock issues.
Whenever terraform starts a state operation (like plan or apply) it acquires the remote lock so that no other terraform execution causes a split brain issue. It would then yield the lock upon the executions completion (successful or otherwise) OR when it received the sigint (e.g. ctrl-C) signal for graceful termination. Funny thing about the subprocess.run function-- it doesn't gracefully handle sigint! If the sigint signal is sent to the python process then it does not wait for the current child process to gracefully handle it, it kills it asap! For really quick process it'd be a race condition between the sigint signal and the sigkill signal just a few milliseconds behind, but as the terraform lock release takes ~3 seconds it basically guarantees that any sigint on opta would cause the lock to not be released. Hence our constant lock problems.
So!
What I did is I made a slightly modded version of the subprocess.run that actually handles the KeyboardInterrupt/sigint/ctrl-c and does wait for the subprocess to handle the signal gracefully.
Tested it out a few times by rerunning an opta gen  ctrl-c which consistently caused lock issues and the issue is gone!