Added support for docdb (#60)
* Added support for docdb

So in order for docdb to have tls encryption during transit, the guides
https://docs.aws.amazon.com/documentdb/latest/developerguide/security.encryption.ssl.html
https://docs.aws.amazon.com/documentdb/latest/developerguide/connect_programmatically.html
Call for the public keys to be downloaded/present in each container using it. The size of the file (~5kb) is just shy of 5% the capacity of a single k8s configmap value so what's happening is each deployment now has a configmap resource which for now only has 1 key (the rds ca pem) which is put in the /config/rds_ca.pem path (and there's now an envar pointing directly to that). With that in place a user can now have their app use docdb by referencing 4 environment variables: username, password, hostname, and ca_path.

WE NEED TO BE SURE TO TELL FOLKS IN OUR DOCDB MODULE DOCS TO USE THAT ENVAR.

* addressed cr