Fixing the terraform init issue with envs
So this actually was an issue with two bugs which cancelled each other off in the service case, but broke the env case. Ultimately, `terraform state list` needs a terraform init even if you downloaded the statefile (bummer). This is supposed to be done on the first apply loop but since we check it based on module idx being 0 (and for our services we always continue to the last module idx), it never surfaced in service opta yamls. Should now be fixed with no extra terraform inits.