Fixing tls
OK so this is why we were having issues with public-uri false.

Cert manager gives a k8s user the ability to issue ssl certificates based on an issues "[type](https://cert-manager.io/docs/configuration/)" (e.g. ACME, self-signed, custom made certificate authority, vault, etc..). There's a one-time step of setting up the issuer and from there your ingresses can ask (via an annotation) "hey I want a certificate from this issuer I created with cert manager" -- cert manager then reads the ingress spec and figures out what hosts the cert is for and says "ok I'll create a fresh new one and put it where you said you will look".

The problem that we were encountering is that if the public-uri is all (e.g. support all hosts) then we tried to make the ssl cert for a wild card "*" which is not supported. Given that the ssl cert is just used to make our communications between the nginx ingress pods and the cloud load balancer encrypted (needed by grpc but also good security practices) and the load balancers do not care about domain verification, the host is totally inconsequential. The fix is to fallback on some determinist dns name and for consistency we now will use the service-name.namespace as that's how the service is identified within the k8s cluster.

One future change we are preparing for: currently, the certificates used are fully self-signed and do not belong to any CA. This _feels_ to remedy this we are setting up a custom-ca issuer with certmanager so that in the future all certs can be managed via this one CA.