Release 0.15.0 (#410)
* Attempting more graceful usage of git dependency (#393)

Import is not within a try-catch and only imported if the code for its usage is invoked.

* Nicer plan displayer (#391)

* Nicer plan displayer

1. Using tables to show data about changes
2. Only 2 security levels "LOW RISK" and "HIGH RISK". Low risk represents no expected dataloss and minimal, recoverable downtime

* more colors

* typo

* addressing cr

* max node count increase should be benign

* addressing cr

* Preemptive GCP instances. (#394)

* Fix unnecessary logs and other minor issues (#395)

* fix-stuff

* revert-pipfile

* fix-dict

* fix-lint

* fix-lint

* Update nightly.yaml

* Checking b/w 2 versions will now verify no dataloss (#398)

* Feature/dashlocal (#397)

Added --local functionality for Opta local runs.

* Check in ci that pdb is not mentioned anywhere in the code (#401)

* Check in ci that pdb is not mentioned anywhere in the code

* wip

* wip

* Tiny fix for our runx module (#405)

* Jd/fixing displayer (#406)

* plan displayer handle delete

* destroy was not a valid value

* adding deletions to test

* Fixed local yaml quote issue (#404)

* Jd/fixing aws destroy (#407)

* AWS destroy facing issue b/c cli can't clean up auto created sec group

https://github.com/terraform-aws-modules/terraform-aws-vpc/issues/283

* wip

* terraform fmt

* Adding Secondary Gcp NodePool Opta module (#403)

* Adding Secondary Gcp NodePool Opta module

* Update the GCP Node Pool name

* Update gcp-env.yml example and fix terraform lint

* Add IAM Member permissions

* Supporting a list of helm values files (#409)

Also, now we check for relative path and support it

* Updating to linkerd v2.10.2 (#408)

* Updating to linkerd v2.10.2

OK, looks like all of the important work was already taken care of earlier as we skipped the outbound mysql and postgres ports already.

More good news: the linkerd visualization components are now their own separate charts so linkerd resource overhead will drop a lot.

https://linkerd.io/2.10/tasks/upgrading-2.10-ports-and-protocols/#
https://linkerd.io/2.10/tasks/upgrade/#upgrade-notice-stable-2-10-0

* terraform fmt

* Fixing gcp dns delegation

* Fixed relative path issues for yaml files (#411)

* Disabling ssl for gcp postgres (#415)

Doing this because in order to enable ssl for postgres a user would need to download the ssl CA/key files and include them in all outgoing connections, which makes psql incredibly cumbersome to use. So for now, we won't be supporting it.

* Forgot to update regula to not complain about missing gcp postgres ssl (#416)

* Fixing azure destroy (#414) (#418)

* Fixing azure destroy

Sometime in the past week Azure terraform started failing when trying to destroy the acr key vault key because we disabled purge and yet destroy causes a purge. Found the new toggle to just do a soft delete on purge and confirmed that it worked.

* Disabling regular rule for purge

* Reverting azure provider to version 2.78.0

Honestly, I'm very disappointed with Azure:
https://github.com/Azure/AKS/issues/2584

* Fixed terraform local working dir (#413)

Co-authored-by: Juan Diego Palomino <jdpcbs@gmail.com>
Co-authored-by: Nilesh Sarupriya <sarupriyanilesh@outlook.com>
Co-authored-by: Sachin Agarwal <sachin@bigbitbus.com>