Fix missing request for memory errors
kubectl events was being filled with "Namespace foobar event: missing request for memory errors"
This was due to the fact that the linkerd sidecars did not have resource requests/limitations set making the horizontal pod autoscaler complain.

This pr fixes it by adding the necessary annotations for cpu/memory requests/limits for the sidecar.