Default Values peristence and upgrade (#491)
* Default Values peristence and upgrade

# The Issue

Right now, for user inputs each module has a default value for each of them to use if the user does not include one themselves. The problem comes for when we want to update the default value and whether the update is NEEDED or optional. We've already had issue where for users their deployment got messed up due to an "optional" change on the default value. Here is the proposed fix:

# Proposal

In the registry modules, each user input will have a new field called "force_update_default_counter" (if it's not set, then it defaults to 0) like so:

```yaml
halt: false
environment_module: true
inputs:
.
.
.
  - name: max_nodes
    user_facing: true
    validator: any(str(), int(), required=False)
    description: Max number of nodes to allow via autoscaling
    default: 15
    force_update_default_counter: 1
```

After a successful opta apply we already store a metadata json in the state storage bucket— now we will add the defaults used

```json
{
  "opta_version": "v0.18.1",
  "date": "2021-11-12T17:37:27.737326",
  "defaults": {
    "module_x": [
      { "input_name": "max_nodes", "default": 15, "force_update_default_counter": 1 }
    ]
  },
  "original_spec": "environments:\n  - name: gcp-staging\n    path: \"../environments/gcp-env.yml\"\n    variables:\n      max_containers: 2\nname: blah40\nmodules:\n  - type: k8s-service\n    image: kennethreitz/httpbin\n    min_containers: 2\n    max_containers: \"{vars.max_containers}\"\n    liveness_probe_path: \"/get\"\n    readiness_probe_path: \"/get\"\n    port:\n      http: 80\n    public_uri: \"test-service.{parent.domain}\"\n    healthcheck_path: \"/healthcheck\"\n    env_vars:\n      - name: APPENV\n        value: \"{env}\"\n      - name: CLOUD_PROVIDER\n        value: gcp\n      - name: hello\n        value: world5\n    links:\n      - pg\n      - redis\n    persistent_storage:\n      - path: \"/DESIRED/PATH1\"\n        size: 5 # 5 GB\n      - path: \"/DESIRED/PATH2\"\n        size: 10 # 10 GB\n  - name: pg\n    type: postgres\n    safety: false\n  - name: redis\n    type: redis\n"
}
```

So upon every opta apply, when generating main.tf.json and selecting the defaults, we shall compare these default values found in the existing metadata json (i.e. the ones used in the last successful opta apply), and reconcile them with the existing ones, using the older values unless the force_default_counter has been incremented, in which case we use the new one found in the registry yaml. After the reconciliation we apply just like before and upload the new default values used. Rinse and repeat

On the first run, it will behave like before— e.g. if no data is found in the metadata.json, it shrugs, keeps the show going and runs with the current default values )because there is no stored state to tell it otherwise).

* testfixing

* test fixing

* testfixing

* Update opta/core/local.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/core/gcp.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/core/aws.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/core/cloud_client.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/core/cloud_client.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/core/cloud_client.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* minor cleanup

* testfixing

* Update opta/layer.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* Update opta/module.py

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>

* linting

Co-authored-by: Patrick Fiedler <rabbitfang@users.noreply.github.com>