After handling the real edge cases, we can begin ignoring some events (#563)
Some of the events hold messages that are 99% of the time more confusing than informative, like for the metrics scraping. We know all of our pods have resources set correctly, so the errors of no metrics only occurs when the pod is starting up (could also be that the metrics server is down in the control plane, but I haven't seen that happen yet).