Massively sped up kubectl config operations through caching
OK, so here is the situation:
Everytime we need to setup the kubectl config to use the cluster created in the current environment (and we do this in many places just to be sure it's done well) it used to take an insane amount of time (~2 seconds) because of API requests. However, we realized that this doesn't need to be the case as the data fetched in these requests for aws and gcp is practically idempotent. We're doing expensive requests to get the data which we already have, and furthermore tainting the user's kubeconfig without their knowledge by doing so.

So this is the solution:
We keep kubeconfig cache files for each environment encountered in the hidden ~/.opta directory. We tell all our functions to use the current such file for the kubernetes config to be used in the HTTP requests, and if the user wants thay can add it to the user-managed default ~/.kube/config files via the configure kubectl command