Updating default EKS version to 1.21
Because our defaults are sticky, this will not force any of our customers to go from 1.18 to 1.21, but rather any new kubernetes clusters will come up with this version. In our talk we did not identify any issues with 1.21, and furthermore we've been running it fine on our GCP daily tests, so this should be totally safe.