Deleting EKS cloudwatch log group if pre-existing (#708)
* Deleting EKS cloudwatch log group if pre-existing

Cloudwatch log groups have this nasty habit of reviving themselves if there's ever one more log sent to them. This is why even though terraform deletes  them on opta destroy and we try to delete it again on post delete, they still stuck around. The consensus was to catch the log groups if they stuck around after a destroy and delete them there.

* typo

* Addressing cr

* testfixing

* fixing