Global Accelerator and the reformed cloudfront experience
OK so this is a long one, but here's what's happening:

In Opta prior to this change, we were not handling dns routing to none-k8s-base well. Like if you had cloudfront with k8s-base and the dns module, Opta couldn't point the traffic to cloudfront, and moreover if someone wrote the records themselves, they would be overwritten by external-dns.

As I was writing the global accelerator, I realized that I was going to face the same problems and had to take a holistic approach in stream lining this process. tl;dr read the modified documents, cause now not only is it super straightforward, but along the way I gave first class support to manual overrides so that Opta can easily integrate with pre-existing dns even better.

Oh, and global accelerator works as you'd think, tried it out my self and with nilesh, and worked great.