Setting kube config and context for helm calls (#826)
* Setting kube config and context for helm calls

OK, this is what was happening:
We have this nice python class/package to manage our calls to helm. These calls to helm are executed as subprocess whose outputs/error we capture for parsing/error handling. The problem is that these functions are run as part of the k8s service processor, and never specify what kube config and context to use. As nowadays we try to be mindful of the user's kubeconfig and don't overwrite it/set the default unless explicitly asked via opta configure-kubectl, there's no guarantee that the helm commands would even be using the correct kube config or context.
This is worse in CI/CD jobs where because it's being run for the first time, there is no default kubeconfig to speak of. This also explains why the stop-gap of running configure-kubectl before hand works.

How do we fix it?
We know what kubeconfig we're using and what's the name of the ocntext. So use them.

* testfixing

* testfixing